requirements-dev.txt
behave>=1.2.6
py-trello>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
